requirements.txt
openai
pandas
numpy
python-dotenv
ipykernel
streamlit
langchain
